// types/loading.ts
export interface Loads {
    horizontalLoadX?: number | string;
    horizontalLoadY?: number | string;
    bearingCapacityLoad?: number | string;
    settlementLoad?: number | string;
    effectiveDepthLoad?: number | string;
    swellingPotentialLoad?: number | string;
    foundationPressure?: number | string;
    loadingCase1Min?: number | string;
    loadingCase1Max?: number | string;
    loadingCase1Avg?: number | string;
    loadingCase2Min?: number | string;
    loadingCase2Max?: number | string;
    loadingCase2Avg?: number | string;
    loadingCase3Min?: number | string;
    loadingCase3Max?: number | string;
    loadingCase3Avg?: number | string;
}