// types/foundation.ts
export interface FoundationData {
    foundationWidth?: number | string;
    foundationLength?: number | string;
    foundationDepth?: number | string;
    foundationBaseAngle?: number | string;
    slopeAngle?: number | string;
    foundationType?: number | string;
    foundationSurfaceType?: string;
    foundationArea?: number | string;
    surfaceFrictionCoefficient?: number | string;
}